[김희진] VacationDraft vo추가, 남은 연차/보상휴가 일수 구하는 메서드 로직 완성, VacationRemainService, VacationRemainMapper 작성, 이제 vacationDraft.jsp에 뿌려주는 작업 예정입니다.

+++ edms/src/main/java/com/fit/vo/VacationCategory.java
@@ -4,9 +4,9 @@ import lombok.Data;
 @Data
 public class VacationCategory {
     private Long vacationNo;
-       private String vacationName;
-       private int minDays;
-       private int maxDays;
+       private String vacationName; // 반차, 연차, 보상
+       private double minDays;
+       private double maxDays;
        private String createdate;
        private String updatedate;
 }
