+++ edms/src/main/java/com/fit/vo/VacationHistory.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class VacationHistory {
 	   private int vacationHistoryNo;
        private int empNo;
-       private String vacationName;
-       private String vacationPm;
-       private int vacationDays;
-       private String createdate;
+       private String vacationName; // 반차, 연차, 보상
+       private String vacationPm; // P, M
+       private double vacationDays;
+       private String createdate; // 발생일자
 }
